Cleanup

+++ pages/api/websites/delete/[id].ts
@@ -11,10 +11,10 @@ export default authenticate(async (req, res, user) => {
     .get()
 
   if (!document.exists)
-    return res.status(404).send('Website not found, dumbass')
-  // if (document.get('uid') !== user.uid) {
-  //   return res.status(401).send('You aren\'t the owner, dumbass')
-  // }
+    return res.status(404).send('Website not found, nice guy')
+  if (document.get('uid') !== user.uid) {
+    return res.status(401).send('You aren\'t the owner, nice guy')
+  }
 
   await document.ref.delete()
   return res.status(200).json({})
